requirements.txt
flask
flask-cors
openai
openai-whisper
torch
ffmpeg-python
redis
tiktoken
Flask-Caching
sentence-transformers
numpy
